requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.8.0
beautifulsoup4==4.8.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
configparser==4.0.2
cryptography==2.1.4
cycler==0.10.0
decorator==4.4.1
docker-pycreds==0.4.0
gast==0.2.2
gitdb2==2.0.6
GitPython==3.0.7
google-pasta==0.1.7
gql==0.2.0
graphql-core==1.1
grpcio==1.21.1
h5py==2.9.0
idna==2.6
imageio==2.6.1
imgaug==0.4.0
joblib==0.13.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
networkx==2.4
numpy==1.16.4
nvidia-ml-py3==7.352.0
opencv-python==4.1.0.25
pandas==0.25.0
pathtools==0.1.2
Pillow==6.1.0
promise==2.3
protobuf==3.8.0
psutil==5.6.7
pycrypto==2.6.1
pygobject==3.26.1
pyparsing==2.4.2
python-apt==1.6.4
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.1.2
requests==2.22.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.1
SecretStorage==2.3.1
sentry-sdk==0.14.1
Shapely==1.7.0
shortuuid==0.5.0
six==1.14.0
smmap2==2.0.5
soupsieve==1.9.2
subprocess32==3.5.4
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tqdm==4.33.0
urllib3==1.25.8
wandb==0.8.26
watchdog==0.10.2
Werkzeug==0.15.4
wrapt==1.11.2